refactor(plugin-text): move sliceNodesAfterSelection into document utils
