+++ src/serlo-editor-repo/plugin-text/utils/document.ts
@@ -1,3 +1,5 @@
+import { Editor as SlateEditor, Transforms } from 'slate'
+
 import { StateTypeValueType } from '../../plugin'
 import type { TextEditorState } from '../types'
 
@@ -6,3 +8,26 @@ export const emptyDocumentFactory =
     value: [{ type: 'p', children: [{ text: '' }] }],
     selection: null,
   })
+
+export function sliceNodesAfterSelection(editor: SlateEditor) {
+  if (!editor.selection) return
+
+  // Create a new line at selection
+  SlateEditor.insertBreak(editor)
+
+  const selectionPoint = editor.selection.anchor.path[0]
+  const childrenCount = editor.children.length
+
+  // Save the sliced nodes to a constant
+  const slicedNodes = editor.children.slice(selectionPoint, childrenCount)
+
+  // Remove the sliced nodes from the current Slate instance
+  Transforms.removeNodes(editor, {
+    at: {
+      anchor: { offset: 0, path: [selectionPoint] },
+      focus: { offset: 0, path: [childrenCount] },
+    },
+  })
+
+  return slicedNodes
+}
